fix(Comment): Action buttons
