+++ lib/hooks/useGlobalBlur.js
@@ -1,7 +1,10 @@
 import React from 'react';
 
 export default (ref, callback) => {
-  const wasOutside = e => callback(!ref.current?.contains(e.target));
+  const wasOutside = e => {
+    const currentRef = ref instanceof HTMLElement ? ref : ref?.current;
+    callback(!currentRef?.contains(e.target));
+  };
 
   React.useEffect(() => {
     document.addEventListener('mousedown', wasOutside, false);
